requirements.txt
rich
python-frontmatter
